create list component
