+++ apps/admin-new/src/pages/Proposals.tsx
@@ -4,7 +4,7 @@ import {
   useDaoProposals,
   useProfile,
 } from '@daohaus/moloch-v3-hooks';
-import { ProposalCard } from '@daohaus/moloch-v3-macro-ui';
+import { ProposalCard, ProposalList } from '@daohaus/moloch-v3-macro-ui';
 import { Button, SingleColumnLayout } from '@daohaus/ui';
 import { JSONDisplay } from '../components/JSONDisplay';
 
@@ -14,65 +14,13 @@ export const Proposals = () => {
     address: address as string,
   });
 
-  const { daoId, daoChain } = useCurrentDao();
-  const {
-    proposals,
-    fetchNextPage,
-    filterProposals,
-    hasNextPage,
-    orderProposals,
-    refetch,
-  } = useDaoProposals();
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
 
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const handleFilter = (e: any) => {
-    if (e.target.value === 'all') {
-      filterProposals({
-        dao: daoId,
-      });
-    }
-    if (e.target.value === 'passed') {
-      filterProposals({
-        dao: daoId,
-        passed: true,
-      });
-    }
-  };
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const handleOrder = (e: any) => {
-    if (e.target.value === 'desc') {
-      orderProposals({
-        orderBy: 'createdAt',
-        orderDirection: 'desc',
-      });
-    }
-    if (e.target.value === 'asc') {
-      orderProposals({
-        orderBy: 'createdAt',
-        orderDirection: 'asc',
-      });
-    }
-  };
-
-  if (!daoId || !daoChain) return <>Current DAO not found</>;
 
   return (
     <SingleColumnLayout>
-      {proposals[0] && (
-        <ProposalCard
-          proposal={proposals[0]}
-          daoId={daoId}
-          daoChain={daoChain}
-          refetchProposals={refetch}
-        />
-      )}
-
-      <Button onClick={() => fetchNextPage()} disabled={!hasNextPage}>
-        More
-      </Button>
-      <Button onClick={() => refetch()}>Refetch</Button>
-      <JSONDisplay data={proposals} />
+      <ProposalList />
     </SingleColumnLayout>
   );
 };
