+++ libs/moloch-v3-macro-ui/src/components/ProposalList/index.ts
@@ -1,0 +1,1 @@
+export * from './ProposalList';
