+++ libs/moloch-v3-macro-ui/src/components/index.ts
@@ -5,3 +5,4 @@ export * from './ProposalHistory';
 export * from './SafeCard';
 export * from './SafesList';
 export * from './ProposalCard';
+export * from './ProposalList';
